Add files via upload
By the far simplest CPU Scheduling algorithm is the First Come, First Served (FCFS) scheduling algorithm.
With this scheme, the process that requests the CPU first, is allocated the CPU first. The implementation of FCFS policy is easily managed with a FIFO queue.
When a process enters the ready queue, its PCB is linked onto the tail of the queue.
When the CPU is free, it is allocated to the process at the head of the queue.
The running process is then removed from the queue. The code for FCFS scheduling is simple to write and understand.

 
We have to find out Completion Time (CT), Turn Around Time (TAT), Waiting Time (WT):
Arrival Time (AT): -Time at which process arrives at the ready queue.
Process ID(PID): - It is the unique number for each process.
Burst Time (BT): - The time required by a process for CPU execution.
Completion Time (CT): - Time at which process completes its execution.
CT = BT+CT

Turn Around Time (TAT): - The amount of Time Taken to complete the process.
TAT = CT-AT
Waiting Time (WT):- The amount of time spent ready to run but not running.
WT = TAT-BT